Issue #35: Fix accidential dropping of cronjobs
If two cronjobs are scheduled at exactly the same point in time, they
both will be inserted into a hashmap with the same key (which is the
next execution time) which causes the first inserted job to be dropped
and never scheduled again. Fix this by inserting the cronjob with a
nanosecond value which is not contained in the map yet.