Make system tests run locally in podman
podman is a much better alternative. Remove manual user handling which
was added to make sure mounted sources from the host keep the original
UID. This is no longer necessary in podman.

Let containers run privileged locally as this is needed to test
capability handling by cinit. Change the system tests for capabilities
for the root process to only check for at least these privileges as
there may be more we are not interested in.