Test for gap days in the future
The code to skip gaps runs before the date is set to the correct value.
Thus it misses cases where the invalid date is not reached by only
setting hour and minute. The current code doesn't crash in this case,
but doesn't act like specified, which is to run at the next possible,
valid time, which is 03:00, and not 03:30 as is currently returned.

However, this test passes and fails to reveal the problem.